spec: add build dependencies that are now required for build
Recent changes refactored a code in a whey that running setup.py
requires more python modules and we need to reflect that in our
spec file in order to build RPMs.

We should not need all of these as they are not used but that will
require a lot of code changes.  We should eventually fix it.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>