virtManager: disambiguize LXC
LXC can be many things and in particular when seeing just "LXC" in the
virt-manager UI one might think of either
- Libvirt-LXC => https://libvirt.org/drvlxc.html
- Linuxcontainers LXC => https://linuxcontainers.org

Especially with the "Linux Containers" that follows in brackets that is
misleading. Lets get rid of this ambiguity and name the backend clearly
to be the former and not the latter.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>