createpool: Remove pool-created callback
The intent is this would select the newly created pool in the
UI after the user creates it. With async events though this
is racy and doesn't tend to work as expected.

It's fixable but it would likely require hanging around for
a period of time waiting for the signal to fire, which
takes more work. Just drop it

Signed-off-by: Cole Robinson <crobinso@redhat.com>