Use InputsChannel instead of DisplayChannel when video is "none"
When Video is "none", SPICE never opens the DisplayChannel.
Currently the code just waits forever for the connection, this
means it's impossible to use SPICE input/audio when using
GPU-passthrough.

Instead of waiting for DisplayChannel, use InputsChannel instead.
This works, but doesn't provide the width and height so test for
the DisplayChannel rather than hitting an exception.