tests: force an English UTF-8 locale
Various checks in the test suite try to match for error/status messages,
so using a localized libvirt (or even using translations of an installed
version of virt-manager in the system) makes those check fail.
Hence, force (using $LANG, and unsetting $LANGUAGE) an English UTF-8
locale.

Thanks Cole Robinson for spotting the right place where to inject the
locale settings.

Fixes: #199

Signed-off-by: Pino Toscano <ptoscano@redhat.com>