virt-manager: implement undecorated window mode
I have a 4k monitor and I have a VM that uses 1920x1080 resolution, which
I would like to lay out in the top-right corner of my monitor.

The problem is that as long as the window decoration, menubar and toolbar
is visible, 1920x1080 does not fit in the 1/4th of my monitor.

Also, the current virtual VGA driver for Windows does not support
non-standard resolutions like VirtualBox, thus I am either seeing a lot of
"black" space around my virtual desktop, or I would have to increase
the size of my virt-manager window.

None of this is appealing, so this patch implements a "windowed" full screen
mode, which is basically an undecorated window, without a menu or toolbar.

The same overlay is available at the top of the window than what is available
in full screen.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>