details: expose dies setting in CPU topology
The dies setting is valid for x86_64 VMs in recent QEMU. This needs to
be exposed in the CPU topology details, otherwise modifications will
potentially loose information.

Although libvirt will report 'dies' for all machines, it is only valid
to set it to a value > 1 for x86_64 VMs today. So the UI option is set
read-only for non-x86_64 VMs unless the XML config already has a value
greater than 1.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>